requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.4
psycopg2-binary==2.9.4
sqlparse==0.4.2
